+++ ContentPatcher/Framework/Migrations/Migration_1_10.cs
@@ -22,6 +22,8 @@ namespace ContentPatcher.Framework.Migrations
             this.AddedTokens = new InvariantHashSet
             {
                 ConditionType.IsJojaMartComplete.ToString(),
+                ConditionType.HavingChild.ToString(),
+                ConditionType.Pregnant.ToString(),
                 ConditionType.Random.ToString()
             };
         }